Model resize and changed room layer
OVRCameraRig now is correctly placed in front of the avatar head. Furthermore, Avatar mesh resizes based on the player eye level.

Besides that, Room colliders' layers are now dynamically being setted on the right layer.

Misses: Room Mesh be placed on the floor level and not on the player height level